add finalizing touches to login and reg buttons

+++ group_project_app/views.py
@@ -81,6 +81,8 @@ def search(request):
     return render(request, 'search.html', context)
 
 def favorite(request, place_id):
+    if request.method != 'POST' or 'id' not in request.session:
+        return redirect('/')
     this_user = User.objects.get(id=request.session['id'])
     client = GoogleMapsClient()
     location = client.detail(place_id=place_id)['result']
@@ -93,6 +95,8 @@ def favorite(request, place_id):
     return redirect('/main')
 
 def account(request):
+    if 'id' not in request.session:
+        return redirect('/')
     context = {
         'user': User.objects.get(id=request.session['id'])
     }
@@ -119,6 +123,8 @@ def update(request):
         return redirect('/account')
 
 def delete_fav(request, business_id):
+    if request.method != 'POST' or 'id' not in request.session:
+        return redirect('/')
     business_to_delete = Business.objects.get(id = business_id)
     business_to_delete.delete()
     return redirect('/main')
